Fix Incorrect calculation of discounts if cost is less than 1USD (#2891)

+++ app/models/discount_code.py
@@ -11,7 +11,7 @@ class DiscountCode(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     code = db.Column(db.String)
-    value = db.Column(db.Integer)
+    value = db.Column(db.Float)
     type = db.Column(db.String)
     is_active = db.Column(db.Boolean)
     tickets_number = db.Column(db.Integer)  # For event level discount this holds the max. uses
